Add test for day 2 part 1.

+++ 2025/day02/day02.cpp
@@ -1,12 +1,12 @@
 #include "day02.h"
 
 namespace day02 {
-    int part1(const std::vector<std::string>& rows) {
+    long part1(const std::vector<std::string>& rows) {
         (void)rows;
         return -1;
     }
 
-    int part2(const std::vector<std::string>& rows) {
+    long part2(const std::vector<std::string>& rows) {
         (void)rows;
         return -1;
     }
